+++ exchange/src/binance/spot/parser/market.rs
@@ -5,15 +5,15 @@ use serde_json::Value;
 
 #[derive(Debug, Deserialize)]
 pub struct KlineDataRaw(
-    u128,    // open time
+    u64,    // open time
     Decimal, // open price
     Decimal, // high price
     Decimal, // low price
     Decimal, // close price
     Decimal, // volume
-    u128,    // close time
+    u64,    // close time
     Decimal, // quote volume
-    u128,    // number of trades
+    u64,    // number of trades
     Decimal, // taker buy volume
     Decimal, // taker buy quote
     Value,   // ignore this field
@@ -47,11 +47,11 @@ pub fn parse_klines(symbol: String, data: &str) -> Result<Vec<KlineData>, serde_
 #[derive(Debug, Deserialize)]
 pub struct DepthUpdateRaw {
     #[serde(rename = "E")]
-    timestamp: u128, // 事件时间
+    timestamp: u64, // 事件时间
     #[serde(rename = "U")]
-    first_update_id: u128, // 首次更新ID
+    first_update_id: u64, // 首次更新ID
     #[serde(rename = "u")]
-    last_update_id: u128, // 最后更新ID
+    last_update_id: u64, // 最后更新ID
     #[serde(rename = "b")]
     bids: Vec<Vec<Value>>, // 买方深度
     #[serde(rename = "a")]
@@ -103,17 +103,17 @@ pub fn parse_depth_update(symbol: String, data: &str) -> Result<DepthUpdate, ser
 #[derive(Debug, Deserialize)]
 pub struct AggTradeRaw {
     #[serde(rename = "a")]
-    agg_trade_id: u128, // 归集交易ID
+    agg_trade_id: u64, // 归集交易ID
     #[serde(rename = "p")]
     price: Decimal, // 成交价格
     #[serde(rename = "q")]
     quantity: Decimal, // 成交数量
     #[serde(rename = "f")]
-    first_trade_id: u128, // 被归集的首个交易ID
+    first_trade_id: u64, // 被归集的首个交易ID
     #[serde(rename = "l")]
-    last_trade_id: u128, // 被归集的末次交易ID
+    last_trade_id: u64, // 被归集的末次交易ID
     #[serde(rename = "T")]
-    timestamp: u128, // 成交时间
+    timestamp: u64, // 成交时间
     #[serde(rename = "m")]
     is_buyer_maker: bool, // 买方是否是做市方
 }
@@ -144,7 +144,7 @@ pub fn parse_agg_trades(symbol: String, data: &str) -> Result<Vec<AggTrade>, ser
 #[derive(Debug, Deserialize)]
 pub struct DepthDataRaw {
     #[serde(rename = "lastUpdateId")]
-    last_update_id: u128,
+    last_update_id: u64,
     #[serde(rename = "bids")]
     bids: Vec<Vec<Value>>,
     #[serde(rename = "asks")]
@@ -196,21 +196,21 @@ pub struct AggTradeStreamRaw {
     #[serde(rename = "e")]
     event_type: String, // 事件类型
     #[serde(rename = "E")]
-    event_time: u128, // 事件时间
+    event_time: u64, // 事件时间
     #[serde(rename = "s")]
     symbol: String, // 交易对
     #[serde(rename = "a")]
-    agg_trade_id: u128, // 归集交易ID
+    agg_trade_id: u64, // 归集交易ID
     #[serde(rename = "p")]
     price: Decimal, // 成交价格
     #[serde(rename = "q")]
     quantity: Decimal, // 成交数量
     #[serde(rename = "f")]
-    first_trade_id: u128, // 被归集的首个交易ID
+    first_trade_id: u64, // 被归集的首个交易ID
     #[serde(rename = "l")]
-    last_trade_id: u128, // 被归集的末次交易ID
+    last_trade_id: u64, // 被归集的末次交易ID
     #[serde(rename = "T")]
-    timestamp: u128, // 成交时间
+    timestamp: u64, // 成交时间
     #[serde(rename = "m")]
     is_buyer_maker: bool, // 买方是否是做市方
 }
